moved cli varinats functionality to clint.py module so that users could use it too
